+++ src/Controller/Traits/LinkSocialTrait.php
@@ -14,6 +14,7 @@ namespace CakeDC\Users\Controller\Traits;
 use CakeDC\Auth\Social\MapUser;
 use CakeDC\Auth\Social\Service\ServiceFactory;
 use Cake\Utility\Hash;
+use CakeDC\Users\Plugin;
 
 /**
  * Ações para "linkar" contas sociais
@@ -31,12 +32,17 @@ trait LinkSocialTrait
      */
     public function linkSocial($alias = null)
     {
-        return $this->redirect(
-            (new ServiceFactory())
+        $authUrl = (new ServiceFactory())
                 ->setRedirectUriField('callbackLinkSocialUri')
                 ->createFromProvider($alias)
-                ->getAuthorizationUrl($this->request)
-        );
+                ->getAuthorizationUrl($this->request);
+
+        $this->dispatchEvent(Plugin::EVENT_BEFORE_SOCIAL_LOGIN_REDIRECT, [
+            'location' => $authUrl,
+            'request' => $this->request,
+        ]);
+
+        return $this->redirect($authUrl);
     }
 
     /**