+++ tests/bootstrap.php
@@ -16,6 +16,12 @@
  * has been installed as a dependency of the plugin, or the plugin is itself
  * installed as a dependency of an application.
  */
+
+use Cake\Cache\Cache;
+use Cake\Core\Configure;
+use Cake\Core\Plugin;
+use Cake\Datasource\ConnectionManager;
+
 $findRoot = function ($root) {
     do {
         $lastRoot = $root;