+++ CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Releases for CakePHP 3
 -------------
+* 8.5.1
+  * Added new `UsersAuthComponent::EVENT_SOCIAL_LOGIN_EXISTING_ACCOUNT`
+* 8.5.0
+  * Added new `UsersAuthComponent::EVENT_BEFORE_SOCIAL_LOGIN_REDIRECT`
+  * Added finder to get existing social account
+  * Improved social login to updated social account when account already exists
+  * Improved URLs in template to avoid issue in prefixed routes
 
 * 8.4.0
   * Rehash password if needed at login